PE: Standalone export - keep .lua files in original state if they cant be parsed. (https://forum.game-guru.com/thread/221339)

+++ GameGuru/ggprecompile.lua
@@ -1,6 +1,7 @@
 function ggprecompile(infile)
 	if infile == nil then return end
 	local p = loadfile(infile)
+	if p == nil then return end -- properly not a lua file, keep original.
 	local outfile = infile
 	local f = io.open(outfile, "wb")
 	f:write(string.dump(p))
